+++ CHANGELOG.md
@@ -5,9 +5,28 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://github.com/global-121/121-platform/compare/v1.74.0...master)
+## [Unreleased](https://github.com/global-121/121-platform/compare/v1.74.3...master)
 
 ---
+## [1.74.3](https://github.com/global-121/121-platform/compare/v1.74.2...v1.74.3) - 2022-12-19
+
+### Changed
+
+- Fix: Added 'id' for 'Export people affected'
+- Fix: Added several 'where' clausules with 'programId'.
+
+## [1.74.2](https://github.com/global-121/121-platform/compare/v1.74.1...v1.74.2) - 2022-12-16
+
+### Changed
+
+- Fix: Look for MAX payment per program for incoming whatsapp message
+- Fix: Handle reminders per program (also with the above MAX fix)
+
+## [1.74.1](https://github.com/global-121/121-platform/compare/v1.74.0...v1.74.1) - 2022-12-15
+
+### Changed
+
+- Fix hack incomming PV whatsapps
 
 ## [1.74.0](https://github.com/global-121/121-platform/compare/v1.73.0...v1.74.0) - 2022-12-14